compute: move logging initialization into its own module
This commit moves the logging initialization code into a submodule under
`compute::logging`. Doing so better separates logging internals from
compute state management and allows us to tighten the interface of the
logging module.
