+++ src/compute/src/compute_state.rs
@@ -12,17 +12,14 @@ use std::num::NonZeroUsize;
 use std::ops::DerefMut;
 use std::rc::Rc;
 use std::sync::Arc;
-use std::time::{Duration, Instant};
 
 use bytesize::ByteSize;
 use differential_dataflow::operators::arrange::arrangement::Arrange;
 use differential_dataflow::trace::TraceReader;
 use differential_dataflow::Collection;
 use timely::communication::Allocate;
-use timely::logging::Logger;
 use timely::order::PartialOrder;
 use timely::progress::frontier::Antichain;
-use timely::progress::reachability::logging::TrackerEvent;
 use timely::worker::Worker as TimelyWorker;
 use tokio::sync::mpsc;
 use tracing::{error, info, span, Level};
@@ -40,7 +37,6 @@ use mz_persist_client::cache::PersistClientCache;
 use mz_repr::{Diff, GlobalId, Row, Timestamp};
 use mz_storage_client::controller::CollectionMetadata;
 use mz_storage_client::types::errors::DataflowError;
-use mz_timely_util::activator::RcActivator;
 use mz_timely_util::operator::CollectionExt;
 use mz_timely_util::probe;
 
@@ -334,172 +330,14 @@ impl<'a, A: Allocate> ActiveComputeState<'a, A> {
     }
 
     /// Initializes timely dataflow logging and publishes as a view.
-    pub fn initialize_logging(&mut self, logging: &LoggingConfig) {
+    pub fn initialize_logging(&mut self, config: &LoggingConfig) {
         if self.compute_state.compute_logger.is_some() {
             panic!("dataflow server has already initialized logging");
         }
 
-        use crate::logging::BatchLogger;
-        use timely::dataflow::operators::capture::event::link::EventLink;
+        let (logger, traces) = logging::initialize(self.timely_worker, config);
 
-        let interval = std::cmp::max(1, logging.interval.as_millis())
-            .try_into()
-            .expect("must fit");
-
-        // Track time relative to the Unix epoch, rather than when the server
-        // started, so that the logging sources can be joined with tables and
-        // other real time sources for semi-sensible results.
-        let now = Instant::now();
-        let start_offset = std::time::SystemTime::now()
-            .duration_since(std::time::SystemTime::UNIX_EPOCH)
-            .expect("Failed to get duration since Unix epoch");
-
-        // Establish loggers first, so we can either log the logging or not, as we like.
-        let t_linked = std::rc::Rc::new(EventLink::new());
-        let mut t_logger = BatchLogger::new(Rc::clone(&t_linked), interval);
-        let r_linked = std::rc::Rc::new(EventLink::new());
-        let mut r_logger = BatchLogger::new(Rc::clone(&r_linked), interval);
-        let d_linked = std::rc::Rc::new(EventLink::new());
-        let mut d_logger = BatchLogger::new(Rc::clone(&d_linked), interval);
-        let c_linked = std::rc::Rc::new(EventLink::new());
-        let mut c_logger = BatchLogger::new(Rc::clone(&c_linked), interval);
-
-        let mut t_traces = BTreeMap::new();
-        let mut r_traces = BTreeMap::new();
-        let mut d_traces = BTreeMap::new();
-        let mut c_traces = BTreeMap::new();
-
-        let activate_after = 128;
-
-        let t_activator = RcActivator::new("t_activator".into(), activate_after);
-        let r_activator = RcActivator::new("r_activator".into(), activate_after);
-        let d_activator = RcActivator::new("d_activator".into(), activate_after);
-        let c_activator = RcActivator::new("c_activator".into(), activate_after);
-
-        if !logging.log_logging {
-            // Construct logging dataflows and endpoints before registering any.
-            t_traces.extend(logging::timely::construct(
-                self.timely_worker,
-                logging,
-                Rc::clone(&t_linked),
-                t_activator.clone(),
-            ));
-            r_traces.extend(logging::reachability::construct(
-                self.timely_worker,
-                logging,
-                Rc::clone(&r_linked),
-                r_activator.clone(),
-            ));
-            d_traces.extend(logging::differential::construct(
-                self.timely_worker,
-                logging,
-                Rc::clone(&d_linked),
-                d_activator.clone(),
-            ));
-            c_traces.extend(logging::compute::construct(
-                self.timely_worker,
-                logging,
-                Rc::clone(&c_linked),
-                c_activator.clone(),
-            ));
-        }
-
-        // Register each logger endpoint.
-        let activator = t_activator.clone();
-        self.timely_worker.log_register().insert_logger(
-            "timely",
-            Logger::new(
-                now,
-                start_offset,
-                self.timely_worker.index(),
-                move |time, data| {
-                    t_logger.publish_batch(time, data);
-                    activator.activate();
-                },
-            ),
-        );
-
-        let activator = r_activator.clone();
-        self.timely_worker.log_register().insert_logger(
-            "timely/reachability",
-            Logger::new(
-                now,
-                start_offset,
-                self.timely_worker.index(),
-                move |time, data: &mut Vec<(Duration, usize, TrackerEvent)>| {
-                    let mut converted_updates = Vec::new();
-                    for event in data.drain(..) {
-                        match event.2 {
-                            TrackerEvent::SourceUpdate(update) => {
-                                let massaged: Vec<_> = update
-                                    .updates
-                                    .iter()
-                                    .map(|(node, port, time, diff)| {
-                                        let ts = time.as_any().downcast_ref::<Timestamp>().copied();
-                                        let is_source = true;
-                                        (*node, *port, is_source, ts, *diff)
-                                    })
-                                    .collect();
-
-                                converted_updates.push((
-                                    event.0,
-                                    event.1,
-                                    (update.tracker_id, massaged),
-                                ));
-                            }
-                            TrackerEvent::TargetUpdate(update) => {
-                                let massaged: Vec<_> = update
-                                    .updates
-                                    .iter()
-                                    .map(|(node, port, time, diff)| {
-                                        let ts = time.as_any().downcast_ref::<Timestamp>().copied();
-                                        let is_source = false;
-                                        (*node, *port, is_source, ts, *diff)
-                                    })
-                                    .collect();
-
-                                converted_updates.push((
-                                    event.0,
-                                    event.1,
-                                    (update.tracker_id, massaged),
-                                ));
-                            }
-                        }
-                    }
-                    r_logger.publish_batch(time, &mut converted_updates);
-                    activator.activate();
-                },
-            ),
-        );
-
-        let activator = d_activator.clone();
-        self.timely_worker.log_register().insert_logger(
-            "differential/arrange",
-            Logger::new(
-                now,
-                start_offset,
-                self.timely_worker.index(),
-                move |time, data| {
-                    d_logger.publish_batch(time, data);
-                    activator.activate();
-                },
-            ),
-        );
-
-        let activator = c_activator.clone();
-        self.timely_worker.log_register().insert_logger(
-            "materialize/compute",
-            Logger::new(
-                now,
-                start_offset,
-                self.timely_worker.index(),
-                move |time, data| {
-                    c_logger.publish_batch(time, data);
-                    activator.activate();
-                },
-            ),
-        );
-
+        // TODO: Construct error collection and trace bundles inside `logging::initialize`.
         let errs = self
             .timely_worker
             .dataflow_named("Dataflow: logging", |scope| {
@@ -508,69 +346,16 @@ impl<'a, A: Allocate> ActiveComputeState<'a, A> {
                     .trace
             });
 
-        let logger = self
-            .timely_worker
-            .log_register()
-            .get("materialize/compute")
-            .unwrap();
-
-        if logging.log_logging {
-            // Create log processing dataflows after registering logging so we can log the
-            // logging.
-            t_traces.extend(logging::timely::construct(
-                self.timely_worker,
-                logging,
-                t_linked,
-                t_activator,
-            ));
-            r_traces.extend(logging::reachability::construct(
-                self.timely_worker,
-                logging,
-                r_linked,
-                r_activator,
-            ));
-            d_traces.extend(logging::differential::construct(
-                self.timely_worker,
-                logging,
-                d_linked,
-                d_activator,
-            ));
-            c_traces.extend(logging::compute::construct(
-                self.timely_worker,
-                logging,
-                c_linked,
-                c_activator,
-            ));
-        }
-
         // Install traces as maintained indexes
-        for (log, (trace, token)) in t_traces {
-            let id = logging.index_logs[&log];
-            self.compute_state
-                .traces
-                .set(id, TraceBundle::new(trace, errs.clone()).with_drop(token));
-        }
-        for (log, (trace, token)) in r_traces {
-            let id = logging.index_logs[&log];
-            self.compute_state
-                .traces
-                .set(id, TraceBundle::new(trace, errs.clone()).with_drop(token));
-        }
-        for (log, (trace, token)) in d_traces {
-            let id = logging.index_logs[&log];
-            self.compute_state
-                .traces
-                .set(id, TraceBundle::new(trace, errs.clone()).with_drop(token));
-        }
-        for (log, (trace, token)) in c_traces {
-            let id = logging.index_logs[&log];
+        for (log, (trace, token)) in traces {
+            let id = config.index_logs[&log];
             self.compute_state
                 .traces
                 .set(id, TraceBundle::new(trace, errs.clone()).with_drop(token));
         }
 
         // Initialize frontier reporting for all logging indexes.
-        let index_ids = logging.index_logs.values().copied();
+        let index_ids = config.index_logs.values().copied();
         for id in index_ids {
             if let Some(frontier) = self
                 .compute_state
